third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-dom-event-dispatching-svg-in-shadow-subtree.html

<!DOCTYPE html>
<html>
<head>
    <script src="../../../resources/js-test.js"></script>
    <script src="resources/shadow-dom.js"></script>
    <script src="resources/event-dispatching.js"></script>
</head>
<body>
    <p id="description"></p>
    <div id="sandbox"></div>
    <pre id="console"></pre>
    <script>
        var sandbox = document.getElementById('sandbox');

        sandbox.appendChild(
            createDOM('div', {'id': 'top'},
                      createDOM('div', {'id': 'shadow-host'},
                                createShadowRoot({'id': 'shadow-root'}))));
        var svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
        svg.id = 'svg-in-shadow-tree';
        getNodeInComposedTree('shadow-host/').appendChild(svg);

        addEventListeners(['top', 'shadow-host/', 'shadow-host/svg-in-shadow-tree']);
        showSandboxTree();

        moveMouse('shadow-host/svg-in-shadow-tree', 'top');
    </script>
</body>
</html>
