third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-dom-event-dispatching-svg-in-shadow-subtree-expected.txt



Flat Tree will be:
DIV	 id=sandbox
	DIV	 id=top
		DIV	 id=shadow-host
			svg	 id=svg-in-shadow-tree class=[object SVGAnimatedString]


Moving mouse from shadow-host/svg-in-shadow-tree to top

  mouseout
     @svg-in-shadow-tree (target: svg-in-shadow-tree) (related: top)
     @shadow-root (target: svg-in-shadow-tree) (related: top)
     @top (target: shadow-host) (related: top)

  mouseover
     @top (target: top) (related: shadow-host)
PASS successfullyParsed is true

TEST COMPLETE
